+++ R/harmonize_mzs.R
@@ -12,14 +12,11 @@ harmonize_mzs <- function(spectra, dalton, ppm) {
 
   # Averaged m/z values from combined spectra
   averaged_intensities <- spectra_new |>
-    Spectra::peaksData(
-      BPPARAM = BiocParallel::SerialParam()
-    ) |>
+    Spectra::peaksData() |>
     Spectra::combinePeaksData(
       tolerance = dalton,
       ppm = ppm,
-      peaks = "union",
-      BPPARAM = BiocParallel::SerialParam()
+      peaks = "union"
     ) |>
     data.frame() |>
     tidytable::pull("mz")